test: Move TLS test data gen into a test package
This will allow it to be used in other packages' unit tests

https://github.com/ollama/ollama/issues/3203
